Use generics for bit depth throughout the encoder
This is a followup from #3116 which expands this optimization
to as many places in the encoder as we can reasonably utilize it.
By using generics, there are places where the compiler is able to
simplify math operations at compile time as well as areas where the
compiler is able to remove branches so that we only branch on bit depth
at the highest level of the code (and therefore the fewest number of
times).

Based on hyperfine benchmarking, this results in a 1-2% speedup across
the encoding process, although it does increase the final binary size.